tests/requirements.txt
coverage>5.2
codecov>=2.1
pytest>=6.0
pytest-cov>2.10
# pytest-xdist
# pytest-flake8
flake8
check-manifest
twine>=3.2
isort>=5.0
mypy>=0.790
yapf>=0.29.0
torchvision>=0.5
piq

cloudpickle>=1.3
pytorch-lightning>=1.0
scikit-learn>0.22.1
scikit-image>0.17.1
nltk>=3.3
